feat: move LetterList into PSR-4 friendly resolver
